添加工资表，未完成

+++ src/main/java/com/lhr13/hpm/dao/PersonDAO.java
@@ -1,6 +1,6 @@
 package com.lhr13.hpm.dao;
 
-import com.lhr13.hpm.bean.Person;
+import com.lhr13.hpm.POJO.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PersonDAO extends JpaRepository<Person, Integer> {
+public interface PersonDAO extends JpaRepository<Person, Long> {
     List<Person> findByName(String name);
 
     @Query(value = "SELECT * FROM personfortest WHERE Dep = ?1", nativeQuery = true)
